refactor: move KsqlSecurityExtension to ksql-engine
This refactor moves the security extension to the
module it deserves to be. It also paves the way to
get user delegation tokens when KSQL executes commands
fetched from the command topic.
