+++ ksql-engine/src/main/java/io/confluent/ksql/security/KsqlSecurityExtension.java
@@ -13,7 +13,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package io.confluent.ksql.rest.server.security;
+package io.confluent.ksql.security;
 
 import io.confluent.ksql.util.KsqlConfig;
 import io.confluent.ksql.util.KsqlException;
